% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{DE_Local}
\alias{DE_Local}
\title{Title}
\usage{
DE_Local(exprsn_contr_column, pExprsn_Location, lfc, p_value, true_gene_list)
}
\arguments{
\item{pExprsn_Location}{a dataframe where starting columns for expression values and finishing columns correspond to gene_id, chromosome and location}

\item{gene_reference}{should be a charactor vector}
}
\value{

}
\description{
Title
}
