% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{DELocal_topTable}
\alias{DELocal_topTable}
\title{Table of Top Genes from DELocal Linear Model Fit}
\usage{
DELocal_topTable(pLinear_model, pDesign_matrix, pLogFold_cut, pValue_cut)
}
\arguments{
\item{pValue_cut}{}
}
\value{

}
\description{
Table of Top Genes from DELocal Linear Model Fit
}
