% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{getCountsFromDESeqData}
\alias{getCountsFromDESeqData}
\title{Title}
\usage{
getCountsFromDESeqData(ddsHTSeq, isNormalized, exprsnColPattern, geneAnnot)
}
\arguments{
\item{geneAnnot}{}
}
\value{

}
\description{
Title
}
