% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{GO_neighbourAnalysis}
\alias{GO_neighbourAnalysis}
\title{Title}
\usage{
GO_neighbourAnalysis(gene_xprsn_annotation, nearest_neighbours)
}
\arguments{
\item{nearest_neighbours}{}
}
\value{

}
\description{
Title
}
