% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{newMakeContrastBvsA}
\alias{newMakeContrastBvsA}
\title{Title}
\usage{
newMakeContrastBvsA(residuals, selectA, selectB, pvalue.fdr, adjustPara, lfc)
}
\arguments{
\item{lfc}{}
}
\value{

}
\description{
Title
}
