% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{optimize_Local}
\alias{optimize_Local}
\title{Optimize DELocal with different number of neighbours}
\usage{
optimize_Local(pSmrExpt, pDesign, pValue_cut, pLogFold_cut, true_gene_list)
}
\arguments{
\item{pSmrExpt}{a dataframe where starting columns for expression values and finishing columns correspond to gene_id, chromosome and location}

\item{pDesign}{a design formula}

\item{pValue_cut}{p-value cut-off value}

\item{pLogFold_cut}{LogFold change cut-off value}

\item{true_gene_list}{should be a character vector}
}
\value{
a result in a form of list
}
\description{
Optimize DELocal with different number of neighbours
}
