% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{performance_analysis}
\alias{performance_analysis}
\title{Title}
\usage{
performance_analysis(Method_name, reference_name, reference)
}
\arguments{
\item{reference}{}
}
\value{

}
\description{
Title
}
