% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{plotNeighbourhood}
\alias{plotNeighbourhood}
\title{Returns median expression from different conditions of genes from a neighbourhood of a gene of interest.}
\usage{
plotNeighbourhood(
  pSmrExpt,
  pNearest_neighbours = 5,
  pDesign = ~condition,
  colorFactor = "condition",
  pGene_id
)
}
\arguments{
\item{pGene_id}{}
}
\value{

}
\description{
Returns median expression from different conditions of genes from a neighbourhood of a gene of interest.
}
