% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{DELocal}
\alias{DELocal}
\title{Finds differentially expressed genes by comparing neighboring genes.}
\usage{
DELocal(
  smrExpt,
  contrast,
  nearest_neighbours,
  pDesign,
  pValue_cut,
  logFold_cut
)
}
\arguments{
\item{smrExpt}{SummarizedExperiment object}

\item{contrast}{Same as DeSeq2}

\item{nearest_neighbours}{How many nearest neighbours within 1 Mb window to evaluate?}
}
\value{
A result table
}
\description{
Finds differentially expressed genes by comparing neighboring genes.
}
