% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{DELocal}
\alias{DELocal}
\title{Finds differentially expressed genes by comparing neighboring genes.}
\usage{
DELocal(pSmrExpt, nearest_neighbours, pDesign, pValue_cut, pLogFold_cut)
}
\arguments{
\item{pSmrExpt}{SummarizedExperiment object}

\item{nearest_neighbours}{How many nearest neighbours within 1 Mb window to evaluate?}

\item{pDesign}{design formula}

\item{pValue_cut}{p-value cut off value}

\item{pLogFold_cut}{Log fold cut off value}
}
\value{
A result table
}
\description{
Finds differentially expressed genes by comparing neighboring genes.
}
