% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{DELocal}
\alias{DELocal}
\title{Finds differentially expressed genes by comparing neighboring genes.}
\usage{
DELocal(gene_xprsn_annotation, sample_names, nearest_neighbours)
}
\arguments{
\item{nearest_neighbours}{}
}
\value{

}
\description{
Finds differentially expressed genes by comparing neighboring genes.
}
