% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalizedLinearModel.r
\name{optimize_Local}
\alias{optimize_Local}
\title{Optimize DELocal with different number of neighbours}
\usage{
optimize_Local(pSmrExpt, pDesign, pValue_cut, pLogFold_cut, true_gene_list)
}
\arguments{
\item{pExprsn_Location}{a dataframe where starting columns for expression values and finishing columns correspond to gene_id, chromosome and location}

\item{gene_reference}{should be a charactor vector}
}
\value{

}
\description{
Optimize DELocal with different number of neighbours
}
